+++ packages/atomic/src/components/facets-v1/atomic-numeric-facet/atomic-numeric-facet.tsx
@@ -109,7 +109,7 @@ export class AtomicNumericFacet
   /**
    * The non-localized label for the facet.
    */
-  @Prop() public label = 'noLabel';
+  @Prop() public label = 'no-label';
   /**
    * The field whose values you want to display in the facet.
    */