+++ packages/atomic/src/components/facets-v1/atomic-timeframe-facet/atomic-timeframe-facet.tsx
@@ -65,7 +65,7 @@ export class AtomicTimeframeFacet
   /**
    * The non-localized label for the facet.
    */
-  @Prop() public label = 'noLabel';
+  @Prop() public label = 'no-label';
   /**
    * The field whose values you want to display in the facet.
    */