+++ packages/atomic/src/components/facets-v1/facet-value-icon-rating/facet-value-icon-rating.tsx
@@ -4,7 +4,11 @@ import {FacetValueIconRatingProps} from '../facet-common';
 export const FacetValueIconRating: FunctionalComponent<FacetValueIconRatingProps> = (
   props
 ) => {
-  const createIconFilled = (active: boolean) => {
+  const width =
+    ((props.numberOfActiveIcons / props.numberOfTotalIcons) * 100).toString() +
+    '%';
+
+  const renderIcon = (active: boolean) => {
     return (
       <div
         innerHTML={props.icon}
@@ -13,33 +17,33 @@ export const FacetValueIconRating: FunctionalComponent<FacetValueIconRatingProps
     );
   };
 
-  const generateIconDisplay = () => {
+  const emptyIconDisplay = () => {
     const emptyIconDisplay: VNode[] = [];
-    const filledIconDisplay: VNode[] = [];
     for (let i = 0; i < props.numberOfTotalIcons; i++) {
-      emptyIconDisplay.push(createIconFilled(false));
-      filledIconDisplay.push(createIconFilled(true));
+      emptyIconDisplay.push(renderIcon(false));
     }
-    const limitString =
-      (
-        (props.numberOfActiveIcons / props.numberOfTotalIcons) *
-        100
-      ).toString() + '%';
-    return (
-      <div class="relative left-0 top-0" part="value-label">
-        <div class="relative left-0 top-0 z-0 flex items-center gap-0.5 pt-0.5 pb-0.5">
-          {emptyIconDisplay}
-        </div>
-        <div
-          id="filledIcons"
-          class="absolute left-0 top-0 z-10 flex items-center gap-0.5 pt-0.5 pb-0.5 overflow-hidden"
-          style={{width: limitString}}
-        >
-          {filledIconDisplay}
-        </div>
-      </div>
-    );
+    return emptyIconDisplay;
   };
 
-  return generateIconDisplay();
+  const filledIconDisplay = () => {
+    const filledIconDisplay: VNode[] = [];
+    for (let i = 0; i < props.numberOfTotalIcons; i++) {
+      filledIconDisplay.push(renderIcon(true));
+    }
+    return filledIconDisplay;
+  };
+
+  return (
+    <div class="relative left-0 top-0" part="value-rating">
+      <div class="relative left-0 top-0 z-0 flex items-center gap-0.5 pt-0.5 pb-0.5">
+        {emptyIconDisplay()}
+      </div>
+      <div
+        class="absolute left-0 top-0 z-10 flex items-center gap-0.5 pt-0.5 pb-0.5 overflow-hidden"
+        style={{width}}
+      >
+        {filledIconDisplay()}
+      </div>
+    </div>
+  );
 };