+++ packages/samples/headless-react/src/pages/SearchPage.tsx
@@ -1,10 +1,7 @@
-import './SearchPage.css';
 import {Component} from 'react';
 import filesize from 'filesize';
 import {AppContext} from '../context/engine';
 
-import {RecommendationList} from '../components/recommendation-list/recommendation-list.class';
-import {RecommendationList as RecommendationListFn} from '../components/recommendation-list/recommendation-list.fn';
 import {Tab} from '../components/tab/tab.class';
 import {Tab as TabFn} from '../components/tab/tab.fn';
 import {BreadcrumbManager} from '../components/breadcrumb-manager/breadcrumb-manager.class';
@@ -44,8 +41,6 @@ import {HistoryManager} from '../components/history-manager/history-manager.clas
 import {HistoryManager as HistoryManagerFn} from '../components/history-manager/history-manager.fn';
 import {RelevanceInspector} from '../components/relevance-inspector/relevance-inspector.class';
 import {RelevanceInspector as RelevanceInspectorFn} from '../components/relevance-inspector/relevance-inspector.fn';
-import {StandaloneSearchBox} from '../components/standalone-search-box/standalone-search-box.class';
-import {StandaloneSearchBox as StandaloneSearchBoxFn} from '../components/standalone-search-box/standalone-search-box.fn';
 import {RedirectionTrigger} from '../components/triggers/redirection-trigger.class';
 import {RedirectionTrigger as RedirectionTriggerFn} from '../components/triggers/redirection-trigger.fn';
 import {QueryTrigger} from '../components/triggers/query-trigger.class';
@@ -107,8 +102,8 @@ import {
   buildHistoryManager,
   RelevanceInspector as HeadlessRelevanceInspector,
   buildRelevanceInspector,
-  StandaloneSearchBox as HeadlessStandaloneSearchBox,
-  buildStandaloneSearchBox,
+  SearchAppState,
+  loadSearchAnalyticsActions,
   RedirectionTrigger as HeadlessRedirectionTrigger,
   buildRedirectionTrigger,
   QueryTrigger as HeadlessQueryTrigger,
@@ -120,16 +115,17 @@ import {
   SmartSnippetQuestionsList as HeadlessSmartSnippetQuestionsList,
   buildSmartSnippetQuestionsList,
 } from '@coveo/headless';
-import {
-  RecommendationEngine,
-  buildRecommendationEngine,
-  getSampleRecommendationEngineConfiguration,
-  buildRecommendationList,
-  RecommendationList as HeadlessRecommendationList,
-} from '@coveo/headless/recommendation';
 import {bindUrlManager} from '../components/url-manager/url-manager';
 import {setContext} from '../components/context/context';
 import {dateRanges} from '../components/date-facet/date-utils';
+
+declare global {
+  interface Window {
+    HEADLESS_STATE?: SearchAppState;
+  }
+}
+
+const isServerSideRendered = !!global.window?.HEADLESS_STATE;
 
 const [KB, MB, GB] = [1e3, 1e6, 1e9];
 
@@ -144,11 +140,13 @@ const initialCriterion = criteria[0][1];
 
 const resultsPerPageOptions = [10, 25, 50, 100];
 
-export class SearchPage extends Component {
-  private readonly engine: SearchEngine;
-  private readonly recommendationEngine: RecommendationEngine;
+export interface SearchPageProps {
+  engine?: SearchEngine;
+}
 
-  private readonly recommendationList: HeadlessRecommendationList;
+export class SearchPage extends Component {
+  private engine!: SearchEngine;
+
   private readonly tabs: {
     all: HeadlessTab;
     messages: HeadlessTab;
@@ -174,7 +172,6 @@ export class SearchPage extends Component {
   private readonly pager: HeadlessPager;
   private readonly historyManager: HeadlessHistoryManager;
   private readonly relevanceInspector: HeadlessRelevanceInspector;
-  private readonly standaloneSearchBox: HeadlessStandaloneSearchBox;
   private readonly redirectionTrigger: HeadlessRedirectionTrigger;
   private readonly queryTrigger: HeadlessQueryTrigger;
   private readonly notifyTrigger: HeadlessNotifyTrigger;
@@ -185,20 +182,10 @@ export class SearchPage extends Component {
 
   private unsubscribeExecuteTrigger!: Unsubscribe;
 
-  constructor(props: {}) {
+  constructor(props: SearchPageProps) {
     super(props);
 
-    this.engine = buildSearchEngine({
-      configuration: getSampleSearchEngineConfiguration(),
-    });
-
-    this.recommendationEngine = buildRecommendationEngine({
-      configuration: getSampleRecommendationEngineConfiguration(),
-    });
-
-    this.recommendationList = buildRecommendationList(
-      this.recommendationEngine
-    );
+    this.initEngine(props);
 
     this.tabs = {
       all: buildTab(this.engine, {
@@ -240,10 +227,13 @@ export class SearchPage extends Component {
     this.facetManager = buildFacetManager(this.engine);
 
     this.geographyFacet = buildCategoryFacet(this.engine, {
-      options: {field: 'geographicalhierarchy'},
+      options: {
+        field: 'geographicalhierarchy',
+        facetId: 'geographicalhierarchy-2',
+      },
     });
     this.objectTypeFacet = buildFacet(this.engine, {
-      options: {field: 'objecttype'},
+      options: {field: 'objecttype', facetId: 'objecttype-2'},
     });
 
     this.fileSizeAutomaticNumericFacet = buildNumericFacet(this.engine, {
@@ -300,10 +290,6 @@ export class SearchPage extends Component {
 
     this.relevanceInspector = buildRelevanceInspector(this.engine);
 
-    this.standaloneSearchBox = buildStandaloneSearchBox(this.engine, {
-      options: {redirectionUrl: 'https://mywebsite.com/search'},
-    });
-
     this.redirectionTrigger = buildRedirectionTrigger(this.engine);
 
     this.queryTrigger = buildQueryTrigger(this.engine);
@@ -315,6 +301,18 @@ export class SearchPage extends Component {
     this.smartSnippetQuestionsList = buildSmartSnippetQuestionsList(
       this.engine
     );
+  }
+
+  initEngine(props: SearchPageProps) {
+    if (props.engine) {
+      this.engine = props.engine;
+      return;
+    }
+
+    this.engine = buildSearchEngine({
+      configuration: getSampleSearchEngineConfiguration(),
+      preloadedState: window.HEADLESS_STATE,
+    });
   }
 
   componentDidMount() {
@@ -340,6 +338,12 @@ export class SearchPage extends Component {
   }
 
   private executeInitialSearch() {
+    if (isServerSideRendered) {
+      const {logInterfaceLoad} = loadSearchAnalyticsActions(this.engine);
+      this.engine.dispatch(logInterfaceLoad());
+      return;
+    }
+
     this.engine.executeFirstSearch();
   }
 
@@ -350,14 +354,6 @@ export class SearchPage extends Component {
   render() {
     return (
       <div>
-        <AppContext.Provider
-          value={{recommendationEngine: this.recommendationEngine}}
-        >
-          <Section title="recommendation-list">
-            <RecommendationList />
-            <RecommendationListFn controller={this.recommendationList} />
-          </Section>
-        </AppContext.Provider>
         <AppContext.Provider value={{engine: this.engine}}>
           <Section title="tabs">
             <nav>
@@ -496,10 +492,6 @@ export class SearchPage extends Component {
           <Section title="relevance-inspector">
             <RelevanceInspector />
             <RelevanceInspectorFn controller={this.relevanceInspector} />
-          </Section>
-          <Section title="standalone-search-box">
-            <StandaloneSearchBox />
-            <StandaloneSearchBoxFn controller={this.standaloneSearchBox} />
           </Section>
           <Section title="redirection-trigger">
             <RedirectionTrigger></RedirectionTrigger>