+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.42.4](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.42.3...@coveo/atomic@0.42.4) (2021-07-21)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+## [0.42.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.42.2...@coveo/atomic@0.42.3) (2021-07-20)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+## [0.42.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.42.1...@coveo/atomic@0.42.2) (2021-07-15)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+## [0.42.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.42.0...@coveo/atomic@0.42.1) (2021-07-15)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [0.42.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.41.0...@coveo/atomic@0.42.0) (2021-07-14)
+
+
+### Features
+
+* **headless:** improve standalone search box experience ([#962](https://github.com/coveo/ui-kit/issues/962)) ([e7c6d22](https://github.com/coveo/ui-kit/commit/e7c6d22d779cdde02d122f05e497de4d6c4ccfea))
+
+
+
+
+
+# [0.41.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.40.2...@coveo/atomic@0.41.0) (2021-07-14)
+
+
+### Features
+
+* **atomic:** atomic rating facet max value prop ([#963](https://github.com/coveo/ui-kit/issues/963)) ([62ddf43](https://github.com/coveo/ui-kit/commit/62ddf436654a7f2cc65e62f7fdc84b17c6a52b2d))
+
+
+
+
+
 ## [0.40.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.40.1...@coveo/atomic@0.40.2) (2021-07-13)